1.8

+++ app/src/main/java/com/example/yiming/hotelmanagment/common/Constants.java
@@ -27,7 +27,7 @@ public class Constants {
     public static final String BOOKED="BOOKED";
 
     //fake information for room
-    public static final int[]  PRRICE={300,330,540,720};
+    public static final Double[]  PRRICE={300.0,330.0,540.0,720.0};
     public static final int[]  BEDS={1,2,3,5};
     public static final int[]  PIC={R.drawable.r1,R.drawable.r2,R.drawable.r3,R.drawable.r4};
